requirements.txt
setuptools
ujson
requests
pandas
pycryptodome